Added more depth to the `Rmd` vignette.
Fixed the `tx_d3alluv` plot.  The indexing in the `lapply` was including the patient IDs as a sequence, which was why things were blowing up.  By increasing the index `+ 1` in the `lapply` call we make sure we start at the first sequence & get a cleaner plot.  This resolves Issue #27.

Removed the redundant call to `cw` in `tx_indiv`, it was causing an error when called as part of the vignette.

Added some description at the top of `reform_nodes`, changed periods to underscores.  It made things look a bit cleaner.