Changed the package build so that it:
1. builds as `txvis`, not `TxVis`.
2. has proper `DESCRIPTION` information for maintainers &cetera.

This means that @sgoring, @johnwilsonICON & @ellenkorol need to make a decision about who will provide long term maintenance for the package.  I've left myself in charge for now, but again, this should be considered.

Also, author order (see the `Author` and `Author@R` lines in the `DESCRIPTION` file) needs to be decided.